TCPMv2: Power on USB device when chipset goes to ON state
Enable power on USB-C ports that are in Attached.SRC state when chipset
goes to ON state. This was implemented as reaction for power state
change event.

This was found when running master branch on nocturne. While performing
shutdown on chromebook all USB-C ports enters Unattached.SNK, due to
reaction on PD_DRP_FORCE_SINK DRP state and PD_EVENT_UPDATE_DUAL_ROLE.
From Unattached.SNK state port goes to Attached.SRC again through
Unattached.SRC and then AttachWait.SRC. Attached.SRC state enables
power on USB-C port, but in this case it is not actually enabled because
chipset is in CHIPSET_STATE_ANY_OFF.

While booting, device port is still in Attached.SRC state, but with
disabled power on USB-C port. As a result USB-C device needs to be
replugged in order to get it working again.

BUG=b:151301058
BRANCH=none
TEST=Run ChromeOS.
     Plug some USB-C device and make sure device is working.
     Shutdown and power on board.
     Check if USB-C device is working. Do not replug it.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I049ae74b48eb2d11538375d48b3a4c87d3dcc291