Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> This is also working on waddledoo.

Zork needs the AutoDischargeDisconnect to be set much earlier than the code that was being used for waddledoo.  So in this churn I moved the AutoDischargeDisconnect as late as it would allow before zork stopped working and enabled that for both.  If this works then we have a solution otherwise will have to setup different calls for the different tcpci implementations.

Patch-set: 4