PD: Move pd_discovery init out of dfp_consume_identity
Moves initialization of pd_discovery from dfp_consume_identity in order
to support folding the cable identity into the pd_discovery structure
(the cable discovery may or may not precede partner discovery).

Also removes a cable structure reset which would clear the cable during
a power role swap, and unnecessary references to the TCPMv1 message ID
trackers which are a part of the PRL in TCPMv2.

BRANCH=None
BUG=b:152417977
TEST=on kindred with TCPMv2, ensure discovery information is cleared
when unpluggging discovered partner and inserting one with no discover
identity response

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I646b3796704fb6980c64d1b9c60327abee19e371