Update patch set 16
Patch Set 16:

(4 comments)

> Patch Set 16:
> 
> > Patch Set 15:
> > 
> > > Patch Set 15:
> > > 
> > > Matthew - where's the proper place for the license in the file accelgyro_bmi260_config_tbin.c (chromius-overlay/master/licenses)?
> > > For context, the license (BSD 3-clause) was provided by Bosch on b:130415695.
> > 
> > accelgyro_bmi260_config_tbin.c in patchset 15 says "Copyright 2020 The Chromium OS Authors."  Is that correct i.e. did Bosch assign copyright to "The Chromium OS Authors?"
> > 
> > If that's correct, i.e. copyright and license are truly identical to the main EC codebase, then no need for a separate license file.
> > 
> > Otherwise, it looks like for EC codebase the typical way to handle code with a different license is to make a third_party subdirectory:
> > https://chromium.googlesource.com/chromiumos/platform/ec/+/refs/heads/master/third_party/
> > 
> > I suggest following that same pattern.  Fix the header comment for any files not copyright The Chromium OS Authors, move them to a third_party subdirectory, and place a LICENSE file there.
> 
> I think that I should follow the pattern that move the file to third_party/. 
> I uploaded the patch set 16 to solve this. Please check whether it is right.

Next step is, how do we get this license contents included in the actual build?

If this were a normal Chrome OS package, we could just reference the license file from the ebuild.  But EC codebase is all one package.

Looking through src/platform/ec/third_party/ more, I see all the subdirectories have a METADATA file.  That's probably a requirement, and may be the only missing piece left.

If adding METADATA makes this license show up in chrome://credits then that's definitely enough.  If it does not show up in chrome://credits, we'll need to loop in either someone more familiar with EC codebase packaging, and/or vapier@.

Patch-set: 16