Update patch set 2
Patch Set 2:

(1 comment)

> The variations all look the same. Why is this being sorted into board level instead of variant?

I assume this is in preparation for a follow-on change to add board specific thermal setting.
This looks ok so far for that, but would be helpful to see the board specific change on top of this one, so we can review together.

Patch-set: 2
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>