speed up AP State polling when CCD is active.
It turns out that the Cr50 not shutting off the AP UART quickly enough
causes violation of the JSL power sequence requirements due to the
leakage caused by the active UART lines.

Let's speed up AP state polling when CCD is active. When CCD is not
active the UART is shut of to start with.

BUG=b:152446186
TEST=running the new image verified that that timing constraints are
    not violated any more. Also verified reliable UART=>USB bridging
    operation.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Iffb7f8bc33c4516bb7cf5cbf58c5ced277cd1aec