ec: Reduced stack usage of gpioget console command
The gpioget console command has a ccprintf() function with over a dozen
parameters. This has excessive stack usage and can trigger stack
overflows. Split the 11 GPIO flags into unique calls with a loop.
This reduces the number of parameters that are passed and handled by
the ccprintf() at once and reduces stack usage.

BUG=chromium:1056780
BRANCH=none
TEST=Connected servod to servo_v4
looped the command 'dut-control servo_v4_uart_cmd:gpioget' 1000 times
TEST=Manual connection to device in the console and executed gpioget,
verified that the format is identical to before.
TEST=make buildall -j

Change-Id: Ied98bab92d68ec248d626375023da4547b052963