fingerprint: Fix non-private sensor init
You do not need the private lib to communicate with the sensor
and fetch the version. You do, however, need to enable
GPIO_DIVIDER_HIGHSIDE on the dev board to communicate with the
sensor. The key part of this change does just that.

Yes, this is somewhat abusing the fpc_private.c naming convention,
but I feel that this is an important enough feature.
If needed, we could refactor to pull out version probing.

BRANCH=none
BUG=b:151190204
TEST=# Disable private repo
     make proj-bloonchipper
     # Flash bloonchipper dev board
     > fpinfo
     # model should not be 0

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: I7e9ab5cc5c6eddaa45e00983f4a941957474c67a