common/vivaldi: Revert Vivaldi changes
This reverts the current vivaldi code in favor of the soon-to-be-merged
new vivaldi framework. The current code is enabled on demand as it hides
behind CONFIG_VIVALDI. In its new form, vivaldi shall always be
enabled for all boards, as explained in go/vivaldi-fw-design.

Planning to land new code for Vivaldi in EC, so might be better to
start with clean slate.

BUG=b:146501925
TEST=Build
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I1b88d069a265ce59c2c36cf534b740a852745b67