Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can we wait until b:153111766 is resolved before continuing with this review?

No, I don't expect it to be resolved soon. The thing her is that we shouldn't be mixing two things:

(1) The current CLs enable Vivaldi keyboard, which basically means:
      - EC will now send action codes and function codes as appropriate for the keyboard (,ost chromeos keyboard = action codes).
       - The top row can have up to 15 keys
       - The top row keys can be re-ordered.

(2) The bug b:153111766 is about a possible new kernel requirement that kernel should be able to determine capability from the keyboard. Note that it does not happen today (today the kernel just assumes all keys are present). So it cannot be holding this review

Patch-set: 2