charger: Add chgnum arg to charger_get_input_current
With the advent of OCPC, one charger per Type-C, it will be required
to retrieve the input current per charger IC.  This commit adds the
chgnum argument to charger_get_input_current.  For boards with a
single charger IC, they should simply pass in 0 for this argument.

BUG=b:147440290,b:148980034
BRANCH=None
TEST=`make -j buildall`
TEST=With other code, verify that queries for input current is
targeted at the correct charger IC.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Iac80255faa539a7b4cfeb495aaed2bf12e62f182