servo_updater: Add Servo V4.1 to flash and servo updater
Adds the Servo V4.1 to the servo updater as the target 'servo_v41'.
Some minor cleanup was done on the servo_updater logic to make it clearer
which issue is occurring and to address common substrings in 'servo_v4' and
'servo_v41' by extracting the binary name from the config files directly
rather than inferring from the name.

BUG=b:153391164
TEST=sudo emerge chromeos-base/ec-devutils
TEST=flash_ec --board=servo_v41 --image ec.bin
  Verify the flash_ec using the STM32 DFU to flash the image on reworked board
  with compatable BC1.2 change
TEST=sudo servo_updater -b servo_v41
  Verify the servo_updater finds the latest servo_v4.1 version and flashes it
TEST=Attempted to flash V4 and V41 devices with each other's images
  verified that the checks prevent flashing the other version.

Change-Id: I77d94ce9a641f39b7ee446af05e52441b7762c96