Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 6: Code-Review+2
> 
> (2 comments)
> 
> In response to your earlier comment, the only other reasonable way I can think of to store the discovery information is a struct with named identity_data members for each of SOP, SOP', SOP'', and maybe SOP''' if we were ever going to support that. If all the boards were going to fully support this, that's what I would prefer, but that approach would make the code to support only SOP on smaller boards more complicated and spread it out over a larger area. An array of configurable length is probably the best we can do.

We could go with named structure elements.  The structure would need an ifdef punctuating it to save space, but if we did it well the ifdef'ing would only be part of the structure itself and some access functions.  That has the advantage of not being susceptible to array overflow, but it would be harder if we need to index/iterate the information.  I'll think about it since the upper part of the CL stack is still waiting on review anyway.

Patch-set: 6
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>