Update patch set 17
Patch Set 17:

(3 comments)

> Patch Set 17: Code-Review+1
> 
> (1 comment)
> 
> I want to wait to +2 until we decide whether this driver should remove support for CONFIG_USB_PD_VBUS_DETECT_PPC, which would eliminate any potential race conditions.  If we remove CONFIG_USB_PD_VBUS_DETECT_PPC it can be done as a separate CL.

In my opinion, we should leave it in because technically it works; and we should avoid race conditions with what we have here now (correct me if I'm wrong). We would prefer the TCPC-based VBUS detection because it generates interrupts so doesn't require polling. That said all TCPCs should support VBUS detection nowadays, so I don't think we will ever use the PPC detection with this part.

Patch-set: 17