Makefile: set the value to the compiler flag CHIP_FAMILY in uppercase
We want to use the value of CHIP_FAMILY to define chip family code via
macros. (See the child CL for details.) In convention, a macro should be
defined with uppercase letters.
This CL changes the lowercase value of CHIP_FAMILY to the uppercase.

BUG=none
BRANCH=none
TEST=No error for "make buildall"
TEST="grep -rw --exclude="build.mk" "CHIP_FAMILY" ./"; Make sure
CHIP_FAMILY is not used in the source. So it won't have the side effect.

Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Change-Id: I077d327c609387e87cd1ac128cf8c9974da21c53