common/keyboard_vivaldi: Enable vivaldi by default for all boards
Enable vivaldi for all boards with a default layout which is the
standard chromeos layout. THe default layout means the following keys
(in this order) in the top row:
BACK,
FORWARD,
REFRESH,
FULLSCREEN,
OVERVIEW,
BRIGHTNESS_DOWN,
BRIGHTNESS_UP,
MUTE,
VOL_DOWN,
VOL_UP

It also means no "Fn" key and screenlock key is present. Note that
if a board does not have the screenlock key physically, the omission of
screenlock key is now needed to comply with kernel requirement that only
those keys to be provided which are actually present on the keyboard. So
Any boards that actually do not have the screenlock key need to provide
set_vivaldi_keybd_config() function.

Also, any keyboards that have chosen to use the config for keyboard
customization have been left alone.

[Note: For this to work, a kernel with 3 atkbd patches ending at this
patchset is needed:
https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2174387
These have been currently backported to 4.19 and 5.4 kernels]

BUG=b:146501925
TEST=Build
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I3c3596bd0eaf5241288473c27bcf61ba7da39593