common/keyboard_vivaldi: Enable vivaldi by default for all boards
Enable vivaldi for all boards with a default layout which is the
standard chromeos layout. THe default layout means the following keys
(in this order) in the top row:
BACK,
FORWARD,
REFRESH,
FULLSCREEN,
OVERVIEW,
BRIGHTNESS_DOWN,
BRIGHTNESS_UP,
MUTE,
VOL_DOWN,
VOL_UP

It also means no "Fn" key and *No screenlock key*. The omission of
screenlock key is to comply with kernel requiremnt that only those
keys to be provided which are actually present on the keyboard. So
Any boards that actually have the screenlock key need to provide
set_vivaldi_keybd_config() function.

BUG=b:146501925
TEST=Build
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I3c3596bd0eaf5241288473c27bcf61ba7da39593