TCPMv2: Don't go sleep, when PD device is connected
Set SLEEP_MASK_USB_PD bit in sleepmask when PD capable device is
connected and unset when there are no PD capable devices connected.
This is done as it is in TCPMv1(23ad46d b:35539451)

The issue appeared during tests of TCPMv2 on nocturne. When EC goes
deepsleep, the first character sent to the console wakes EC up,
but the command is not received correctly.

BUG=b:152865530
BRANCH=none
TEST=Check bit SLEEP_MASK_USB_PD(0b100000) in sleepmask with EC command
"sleepmask". The bit should be set when PD device is connected and unset
if isn't

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I99e24537b48a0ee9f252a6dcf2c0383b9228a9aa