Implement sequence for detecting trigger for AP RO verification
When supported, the AP RO verification would be triggered by the
operator pressing and keeping pressed the power button and then
pressing and releasing a few times the escape key.

As proposed in this patch, to trigger the verification the operator
must complete the sequence within 3 seconds by pressing the escape key
three times.

The sequences is controlled by periodic polling. Enabling Esc key
press interrupts was investigated, the issue is that the key generates
plenty of interrupts due to dribbling, to the tune of a hundred each
time it is pressed. It is much cheaper to just poll every 20 ms.

The CONFIG_AP_RO_VERIFICATION config flag controls enabling of this
feature.

BUG=b:141191727
TEST=enabled the new feature and verified proper operation by both
    detecting the trigger and abandoning the sequence due to released
    power button or not enough times pressed Esc key.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I55376a87009d6f8020358ad11db1e47d0b8393ed