it81202/it81302: set GPIOH7 as output low at default
The GPIOH7 does not utilize in IT81202/IT81302 and is left
floating internally.
We need to enable internal pull-down for the pin to prevent
leakage current, but IT81202/IT81302 doesn't have the capability
to pull it down.
We can only set it as output low, so we enable output low for it
at initialization to prevent leakage.

BUG=none
BRANCH=none
TEST=On it81202 EVB, power consumption is reduced by ~70uA.

Signed-off-by: Dino Li <Dino.Li@ite.com.tw>
Change-Id: Iba1f5af684e31cad192b4e9ee5b099e7c51a71a3