gsctool: allow to enable write protection (WP)
The gsctool utility allows to examine the device WP status, but does
not allow to set it. It would be useful to provide the user with a
means of enabling WP at any time.

This patch extends the existing vendor command VENDOR_CC_WP to allow
an optional one byte parameter. If the parameter is present, the Cr50
will unconditionally invoke set_wp_state(1) when processing the
command.

BUG=b:153881773
TEST=verified that attempts to enable WP when running the unmodified
     Cr50 image fail with error message "Early Cr50 versions do not
     support setting WP", and that the updated Cr50 image allows to
     enable WP using 'gsctool -a -w enable'

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I75c200bbb9085e9f74c227ef80f782defdaaa29e