Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Only ITE code base for Ampton. After Modify this EC FW, Apshutdown is OK.

Could you expand on why this is needed for one board and not the others?  And add a comment and/or commit description which explains the issue further (what's going wrong currently, how this fix helps, why it's only needed for ampton)?

It's important to ensure the code is well documented for issues like this.

Patch-set: 2
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>