stm32g4: Clock module driver
This CL adds support to configure rcc module for stm32g4. The driver
from the F4 family was used as a reference. Support for RTC was not
ported as it's not being used for honeybuns.

The function wait_for_ready() was moved to the common clock file for
the F-family so it would not need to be replicated for G4 as well.

BUG=148493929
BRANCH=None
TEST=verfied that the GPIO, clocks, and EC console over LPUART

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I980c8889965a2e5da401ccd6291079a0bdfa8e4f