BB retimer: Refactor Re-timer_Driver bit
Ref: Burnside Bridge spec Table 13: Connection state register
In case the port is DFP, if the mux state is either USB or DP, the
re-timer_Driver bit is set according to the SOP' VDO2 response. If the
mux state is Thunderbolt-Compat, the re-timer_Driver bit is set
according to the Discover mode SOP' response from the cable.

BUG=b:152544514
BRANCH=None
TEST=make set re-timer_Driver bit for active cable.

Signed-off-by: Ayushee <ayushee.shah@intel.com>
Change-Id: Idb1134600617161137c6f500790b6c0c3b836200