Update patch set 4
Patch Set 4: Verified+1

Verified +1.
Data and scope traces on b/147833952 .

Note:
* Some boards are "fine" in C-to-C configuration,
  but experience issues in C-to-A-to-C configuration.
* Other boards are affected in both configurations.

This implies there are some subtle timing issue as well.
Please review scope traces for this observation.

Summary:
Hatch/Kohaku(old): C-to-C ok, C-to-A-to-C affected
Hatch/Kohaku(patched): C-to-C ok, C-to-A-to-C ok

(old -- no patch available yet)
Octopus/Yorp/Sparky360: C-to-C ok, C-to-A-to-C affected
Grunt/Kasumi360: C-to-C affected, C-to-A-to-C affected

Patch-set: 4
Reviewer: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1