stm32g4: Add flash support for stm32g4 chip family
This CL adds support for flashing stm32g4. The L4 family has the same
flash module and so that was used as basis for G4 driver. This CL
modifies the registers file to match the L4 module and the driver
itself is symbolically linked to the L4 driver.

BUG=b:148493929
BRANCH=None
TEST=verfied usb-ep fw update completes successfully

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I6cc818c6329524924534b0566bd9d24145c86c72