USB/PD: Skip pd_prepare_sysjump for EFS2 boards
Since in EFS2, USB/PD won't be enabled in RO or if it's enabled in
RO, EC won't jump to RW, pd_prepare_sysjump is not needed.

Even if PD is enabled because the device is not write protected,
EFS2 jumps to RW before PD tasks start. So, there is no states
to clean up.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:137493121, b:154440914
BRANCH=none
TEST=Boot Puff successfully.

Change-Id: Ie598105779d2a605cb63842cb6c8428f0e2c5367