Update patch set 4
Patch Set 4:

(1 comment)

Looks like this will work as described, however, we should think about consume (eg how long is the line asserted for?) - we don't want to allow multiple assertions for one press.

Is the GPIO_DIOM4 line *always* guaranteed, on every device, to be wired to either the FPMCU or nothing? And do we have a guarantee that this line will only be asserted when there was a match, including on older versions of FPMCU firmware that don't know to signal to cr50 (it was always be false there)?

Also, I thought we were going to only allow this for the 'new' key handles?

Patch-set: 4
Reviewer: Gerrit User 1256783 <1256783@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>