Plug in the AP RO verification implementation
This patch enables the invocation of the AP RO verification function
in response to the operator entering the 'magic sequence' of holding
the power button pressed and pressing/releasing the refresh button
three times within five seconds.

The code used during the 'Open box RMA' verification process is used,
with the physical presence confirmation phase bypassed.

BUG=b:153764696
TEST=generated AP integrity verification data using the ap_ro_hash.py
     script and then ran the verification procedure, observing the
     'hash match' message on the Cr50 console.

     Also verified that the Open Box RMA procedure still succeeds.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ic101fb892554ebb05f9ebe6d1546bfb439f74043