tcpmv2: SNK HardReset cleanup
tc_hard_reset tells tcpc we are starting a hard
reset.  It now disables AutoDischargeDisconnect
and sets the flag to put TC in reset.  When
this happens, TCPC should not go to an
unattached state.

tc_hard_reset_complete was added to tell the
TCPC we are done with the hard reset and going
to an unattached state is allowed again.  It
enables AutoDischargeDisconnect and clears
the flag that blocked the TCPC from going to
an unattached state.

hard reset changes the TCPCI mask register
to be their default values.  Alert will now
detect this and will make sure the TCPC is
initialized and back to what we expect it
to be.

BUG=b:154073581,b:154622375,b:154937258
BUG=b:154778807,b:155132904
BRANCH=none
TEST=verify charger is still detected after hardreset

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I3e1dd04b04653b8fca5dccc9373533d5a5ff9586