Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Which alert was getting lost?
> I would have expected them to stay set until we write them to clear them.

I would think you would be correct but my doubt is in the TCPC_REG_ALERT_RX_STATUS.  It should remain set until it is written to clear it as well.  I don't see a write of this bit and yet we leave the loop.  It feels safer to do it this way. If you disagree, I will drop this CL.

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>