TCPMv2: Set AMS start for messages in pe_attempt_port_discovery
Discovery steps (discovering identity, SVIDs, mode, etc.) represent the
beginning of AMS's, as does initiating data and vconn swaps.  This change
sets the local AMS flag so ready exit will let the PRL know we're
beginning an AMS, and ensure that any PD 3.0 partners know to not
interrupt us.

BRANCH=None
BUG=b:155087579
TEST=on kindred with TCPMv2, PD 3.0 port partner doesn't send messages
in the middle of discovery

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I0763ae71551ac0d2c19e3108fed264ec23a9ecd3