Update patch set 1
Patch Set 1:

Honestly, instead of picking a percentage at random, one should actually measure how much the device is overdrawing and compensate for that. Otherwise, you may leave a little bit of power on the table or it may not be sufficient.

If all hatch devices use the same charger IC, it's probably fine to make the reduction common.

Patch-set: 1
CC: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>