SM5803: Only enable auto charging when battery is present
Enabling auto charging bits when the battery is not present can cause
Vsys to brown out when the charger believes charging is complete.  This
uses the battery presence in the driver as an indicator for whether
these bits should be set when enabling the charger.

BRANCH=None
BUG=None
TEST=on waddledee with no battery and charger connected to C0,
Vsys does not brown out

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I3714261f9394b101ddeb76889ebcecba6b61d456