common: Restrict gpioset console command for FPMCU
Generally, we don't want people to be able to modify GPIOs in
production. The main EC already disables the whole console input
in production. For the FPMCU, the host command version of gpioset
is already restricted. This change restricts the console command
version on FPMCU, without changing other boards' behavior.

BRANCH=none
BUG=b:154655729
TEST=make -j buildall

Change-Id: I9ab893b5742926732148e7c363d8aa7dcd7466f2
Signed-off-by: Yicheng Li <yichengli@chromium.org>