charge manager: only debounce CHG_TYPE with battery
The charge manager always reports USB_CHG_TYPE_UNKNOWN within two
seconds of a port being connected to the system in order to give time
for negotiation to occur. On batteryless systems however, it doesn't
make sense to debounce this- CHG_TYPE_UNKNOWN seems to imply "we're
waiting to see if we can actually charge," but if the system has power
and doesn't have a battery it's obvious that it's receiving power from
somewhere- presumably the reported port.

This fixes an issue on Puff where firmware wants to determine the
available power early in boot; possibly less than two seconds after EC
reset, in which case no matter the active port it's always being
debounced. The debounce is unnecessary because by the time AP firmware
has booted, the EC has already successfully negotiated for power.

BUG=b:151972149
BRANCH=None
TEST=pending

Change-Id: If8168c8fb67710e8e84c055b9f2c19f18751ac27
Signed-off-by: Peter Marheine <pmarheine@chromium.org>