Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is this for FP MCU? Main EC disables the whole console input in production. So, no console commands can be run (unless CCD is open).

This is intended primarily for FPMCU. The FPMCU has the console still enabled in software during production, but does not currently have any easy path to physically connect to the signals.

Does main EC actually disable the console in software, or is it simply disabled by the Cr50 gateway?

Patch-set: 1
CC: Gerrit User 1334530 <1334530@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>