Update patch set 4
Patch Set 4:

(5 comments)

I can upload the changes for the comments given but don't want to delay getting the base of this into the repo.  It is enough that lab will be unblocked but uploading another version would let me also upload a third CL that takes care of the Soft Reset issue that caused this whole Hard Reset on trembyle to happen in the first place.

Patch-set: 4
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>