power: Don't boot after waking up from hibernation
When a device shuts down and is idle for 60 mins (configurable), EC
enters hibernation. When an AC adapter is plugged, the system boots.
This is not expected behavior and not consistent with the behavior
that the system stays off on AC plug-in before hibernation.

This patch fixes the above inconsistency by storing AP_OFF flag before
entering hibernation after 60 mins idle in S5.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chromium:1073960
BRANCH=none
TEST=Verified Bloog stays off after waking up from hibernation on
     AC plug-in.

Change-Id: I097bee97164284dd4c35f8bf9389c76319fd676a