keyboard_8042: Add comment about KEYBOARD_ROW_REFRESH
Add comment about the need to change KEYBOARD_ROW_REFRESH in case
the refresh key is moved from T3. This existed pre-vivaldi and needs
to be a build time switch because it is used by EC early on, such that
boards will not get a chance at runtime to detect SKU ID or fw_config
and take decisions this early.

BUG=b:146501925
TEST=Build
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I23ebf1dd567240d6e0dcf4a60ce75090fc9b597b