system: Add alternative system_is_locked behavior
This CL adds CONFIG_SYSTEM_LOCKED_ON_RO_NOW, which alters the behavior
to system_is_locked to only report unlocked when ro_now is disabled.

The current flash-enabled implementation allows either ro_now or HW WP
to "unlock" the system. The FPMCU is intended to remain in a secured
state until the flash protections have been disabled, which for
bloonchipper is effectively a reboot with HW WP disabled.

BRANCH=none
BUG=b:147521114
TEST=Check console commands in all combinations of HW and SW WP state.

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: Ia1fb4c1684c20b15e4efafe54e3669855fd77524