Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I know we're done with this CL - but did we actually measure for the concerns that Jett-Diana-Abe raised about smaller UART buffers blocking? From patchset #5. 
> 
> also - why are the recommended sizes different from 512 to 1024 to 2048 bytes? Was there any experimentation there?

I didn't attempt to observe how easy it is to make the AP miss console logs. I think in order to do that, I would need to get the affected boards to test with.

The sizes are required to be powers of 2. For each board that was short on space, I reduced its current size by one power of 2, which is the minimum amount of reduction possible via this mechanism.

From what I gather so far, there is not currently a recommended size for this buffer, but each chip has a default buffer size. For any particular board, there is probably a minimum size below which the AP would predictably miss logs and a maximum size above which the EC would predictably waste space. For a given feature set, there is presumably a size in that range that comfortably mitigates the risk of missing logs with an acceptable overhead. I don't know whether it is worth trying to establish general guidelines for buffer sizes, given that the verbosity of logs and the need for spare RAM goes up over time.

Patch-set: 8