Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> nDiscoverIdentityCount only applies in Explicit Contract.

I don't think that's the intent of the counter, given it's referenced in PE_SRC_VDM_Identity_Request State which is 100% definitely running before an explicit contract is in place.  

The timer only applies to while we're in an explicit contract, and I clarified with the WG there's no specific discovery timing requirement before a contract (I just choose to still space out requests, because spamming too quickly seems like a poor plan).

For the plan forward here, I'm going to try searching the PD archives to see if I can find discussion as to what prompted the retry paragraph in the specification revision section and see if it has clarification about circumstances assumed and intent.

Patch-set: 1