Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1:
> 
> (3 comments)
> 
> Okay, testing with waddledoo:
> In S0:
> - charger connection good
> - unpowered Apple dongle connection good
> - powered Apple dongle connection not good - power role swap disconnects
> 
> In S5:
> - charger connection good
> - unpowered Apple dongle doesn't connect (good)
> - powered Apple dongle connection good (no power role swap there)
> 
> Unable to compare the bad powered dongle connection to ToT, as my Doo only boots to S0 once in a blue moon (you should feel honored that it decided to boot up just to test your CL ;) )  I could get out my waddledee to test, but as it neither uses autodischargedisconnect nor auto toggling it wouldn't be a great reference point.

Yea. My change definitely has a lot to do with AutoDischarge and AutoToggle.  Are the results above indicating only one thing did not work as expected?  Can you tell me how to reproduce anything that failed and maybe I can use the servo_v4 or another one of my powered hubs to get it to fail for me and I will try to get that fixed.

Thanks for testing it in the early stages.  This gives me some confidence that things are working pretty good. I am still seeing some faults from TCPCI that I am not expecting but getting out of the current fire seems to be my constant M.O.

Patch-set: 1